Add get for all projects test
